CYB-87: refactored code
